added entire clustering for the clusterGraph so that cluster layout makes sense

+++ Node.hpp
@@ -10,7 +10,7 @@ class Node {
 public:
     // Constructor
     Node() = default;
-        Node(int id, int x, int y, int cluster = -1) : _id(id), _x(x), _y(y) , _cluster(cluster) {}
+        Node(int id, double x, double y, int cluster = -1) : _id(id), _x(x), _y(y) , _cluster(cluster) {}
 
 
     int getCluster() {
@@ -21,11 +21,11 @@ public:
         return _id;
     }
 
-    int getX() const {
+    double getX() const {
         return _x;
     }
 
-    int getY() const {
+    double getY() const {
         return _y;
     }
 
@@ -43,8 +43,8 @@ public:
         std::cout << "Node ID: " << _id << ", x: " << _x << ", y: " << _y << "cluster: " << _cluster << std::endl;
     }
 
-    int _x{ 0 };
-	int _y{ 0 };
+    double _x{ 0 };
+	double _y{ 0 };
 	int _id{ 0 };
 	int _cluster{0};
